pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Set of agent based on langgraph tech stack to solve data compliance."
authors = [
    {name = "Pankaj Sharma",email = "pankajsharma96277@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
packages=[{include="sentinel", from="src"}]
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "langchain-google-genai (>=3.0.3,<4.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "qdrant-client (>=1.17.0,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-core (>=1.2.14,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "google-adk (>=1.25.1,<2.0.0)",
    "fastembed (>=0.7.4,<0.8.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.sentinel.main:start"
dev   = "src.sentinel.adk_dev:launch"

[virtualenvs]
in-project = true